// mobile/src/api/schema.generated.ts
export interface components {
    schemas: {
// ...
        IdentityResponse: {
            identity: components["schemas"]["Identity"];
            profile: components["schemas"]["Profile"] | null;
            /**
             * Resource Version
             * @default 1
             * @constant
             */
            resource_version: 1;
        };
// ...
        NativeAuthExchangePendingResponse: {
            /** Exchange Id */
            exchange_id: string;
            /**
             * Resource Version
             * @default 1
             * @constant
             */
            resource_version: 1;
            /** Retry After Seconds */
            retry_after_seconds: number;
            /**
             * Status
             * @constant
             */
            status: "pending";
        };
        /** NativeAuthExchangeSuccessResponse */
        NativeAuthExchangeSuccessResponse: {
            /** Access Token */
            access_token: string;
            /** Expires At */
            expires_at: number;
            /**
             * Resource Version
             * @default 1
             * @constant
             */
            resource_version: 1;
            /**
             * Status
             * @constant
             */
            status: "authenticated";
        };
        /** NativeAuthStartResponse */
        NativeAuthStartResponse: {
            /** Challenge Id */
            challenge_id: string;
            /** Expires At */
            expires_at: number;
            /**
             * Resource Version
             * @default 1
             * @constant
             */
            resource_version: 1;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}